Double delimiters for emphasis and strong
Require double delimiters for emphasis and strong, i.e., '__' and '**'.
This was changed as '_' and '*' are quite often used in texts and
requiring to always escape them harms readability. Additionally,
newcomers may not know about escaping, making it hard for them to use
underscores and asterisks. Requiring double delimiters usually should
not hurt readability as much.