Parse unordered lists, document guiding principles
Write a draft of lists specification. Parse unordered lists including
nested lists and add a few test cases for lists.

Write guiding principles and goals in the specification.

Rename elements node.String() method to node.Pretty() as it better
reflects the actual method which returns a prettified, indented string
representation of an element, not just a string. That way we leave the
String() method for its more common use case and a method that can be
used in the Stringer interface.