Parse lists without indentation backtracking
Before we used p.indentation() method to count the current indentation.
p.indenation() reset its position after execution (backtracked) and if
we continued we consumed that indentation again.

Now list parslets always consume the ending indentation but return it so the
recursive functions always get the ending indentation... Ending
indentation is the indentation that ended the current list or list item.
For example, if we have a nested list, the nested list parses until the
indentation is less than or equal to  the starting indentation. If equal
indentation, it must be a list item to still continue the list. We
always consume the indentation and when the consumed indentation ends
the list, we return it so the parent list parslet can use that
indentation. If we didn't return it, the parent parselet would have no
way to get the indentation as it was already consumed by the nested
parselet.