Add printer.Pretty() instead of node.Node.Pretty()
Remove the method Pretty() from node.Node interface as it is not
necessary. Add printer.Pretty() function which does the same job. It
takes a node or now even a slice of nodes and an indentation level and
returns a prettified, indented string representation of the node/slice
of nodes.

go fmt the renderer package which I forgot in the previous commit.