Fix '>' causing infinite loop
We got an infinite loop if an '>' was encountered without a preceding
'<' in the same line (indicating a link). This happened because
parseText() parsed until a '>' and then returned to parseInline().
parseInline() does not recognize a '>' delimiter (as it is a closing
delimiter) and called parseText() which in turn called parseInline()
again...

parseText() now always parses until an OPENING inline delimiter which
when returned to parseInline(), parseInline() knows how to handle it.
And when appropriate it parses also until a closing delimiter which is
passed explicitly. E.g. we pass the '>' delimiter to parseText() when
parsing text for link.