Fix mixed list indentation, improve spec
Before, a tab equalled one space if indenting a list. For example, a tab
indented list item would equal a one space indented list item. This
visually does not make sense as the tab indented list item will usually
appear as a nested list item.

Now, a tab equals 4 spaces. This is not a perfect solution as the
problem persist for whenever tabs are presented as less or more than 4
spaces/columns. But, it is a more sensible default. Also, it is now
easier to expose this tab width as an option.

Improve specification:
- require UTF-8 encoded source code
- define whitespace
- list disallowed code points
- improve lists description