Report illegal characters, consistent spacing
Report an error if encountered NUL character, byte order mark, or
invalid UTF-8 encoding. Byte order mark is allowed as the first
character.

An encoding is invalid if it is incorret UTF-8, encodes a rune that is
out of range, or is not the shortest possible UTF-8 encoding for the
value.

Consistent spacing: trim only '\t' and ' ' spacing/whitespace, not
Unicode space that the strings.TrimSpace() trims.

Move constants closer to the usage - unimportant, but did not bother to
make a separate commit...