Playing with a "parsing" scanner
Trying out a scanner that would be parsing; tracking state of open
blocks and generating opening and closing block tokens instead of
generating delimiter tokens. The grammar and the actual parser would be
a lot simpler.

The scanner nor the parser actually work, but a parsing.to document was
drafted that describes the scanner parsing strategy.

Here is an example source of an offside block element (indented):
-a
 b

The "parsing" scanner would generate the following tokens:
BEGINLISTITEM  -
TEXT           a
LINEFEED       \n
TEXT           b
EOF            ""

For comparison, a normal scanner would generally generate:
HYPEN     -
TEXT      a
LINEFEED  \n
INDENT    " "
TEXT      b
EOF       ""