Parse certain continuation walled elements
Currenly, the continuation check advances our pointers to the position
after the block delimiters. Because of this, nested blocks are not being
closed properly.

Example nested blockquote:
	>>a
	>b

In the example above, the first continuation check on line "b" is
performed by the nested blockquote from line "a". This check sets p.ch=b
and closes the nested blockquote. Now, the continuation check by the
parent blockquote starts checking the contiuation from p.ch=b onward
which of course does not match the ">" delimiter.