Handle spacing between block delimiters
Discard spacing (CHARACTER TABULATION and SPACE) in between block
delimiters and at the start of line.

Remove leading spacing in nested fenced elements. Example code block
nested in a blockquote:
	> ``
	> a

In the example above, we remove the space before the "a". The amount of
spacing between the parent delimiter and the opening delimiter of the
fenced block is removed for every fenced element line.