Parse forward inline elements
Forward is an inline element delimited by a single character. Forward
consists of a required, raw part and an optional, nestable part that
must be placed right before the required part. Each part must be
delimited. If a delimiter is a left/right character, the closing
delimiter must be the counterpart characeter.

Forward element is intended to be used instead of the composite element.
This is because the forward element is simpler. There should be few such
elements and as such there should not be many unexpected elements
appearing while just writing.