Fix inlines w/o content; only left/right escape ch
Escaped and forward elements did not check for the closing delimiter if
it was placed right after the opening delimiter—ifthere was no content.
As such, they returned the closing delimiter as content in such cases.

Escaped elements now accept only left/right characters as an escape
character instead of any punctuation character.