Fix inlines incorrectly closing nested elements
Example:
	__**a__b

In the example above, the closing "__" closed the emphasis but not the
nested strong.

Similarly, in the first (inline) part of a two-part forward element, the
parsing of the element did not stop if it was closed by another element.

Example 2:
	|**<**><a>

In the example above, we got "Line(Strong(Link("<a", )))" whereas we
want "Line(Strong(Link("", ))Text(">")Link("a", ))".