Refactor paragraphs
Remove the paragraph element and the paragraph group. Add a paragraph
transformer which works similar to the group transformer.

Line breaks are no longer rendered:
	ab
and
	a
	b
are the same.

Filled lines followed by another block or filled lines separated by
blank lines are grouped as a paragraph.

Examples when paragraph is added:
1.
	a

	b
2.
	a
	>
3.
	-a

	 b

Examples when paragraph is not added:
1.
	a
2.
	-a

This approach was chosen as it is orthogonal as opposed to the previous
one and does not require paragraph delimiters.

1.
	a

2.
	> a

Previously, examples 1 and 2 produced different nodes for 'a'.