Fix fmt inline escape not working across elements
Printer/formatter did not add an inline escape if it was on the last
character in the text. Check for the delimiter of the next inline
element if present to fix this issue.

Example:

Text("a`") Code("...")

Note that Code delimiter is "``".

At "`" character in Text we did not check the combination with the first
delimiter character of code and therefore did not detect the need for
escape.