Break up Hanging into Hanging and RankedHanging
Break up because we differentiate between two elements with the same
delimiter where one of them is ranked and one not. For example,
non-ranked heading is a title and should appear at most once in a
document. On the other hand, ranked headings define structure and can
appear multiple times. The point is they are not the same even though
they seem related.

Additionally, configuration is simpler as we do not need the "ranked"
and "minRank" parameters. There is no possibility to set "ranked" on an
element that does not support ranks.

However, two elements with the same delimiter are now allowed if they
are a combination of a Hanging and a RankedHanging elements. The
RankedHanging element takes precedence when parsing. For example, "=="
is a heading not a title.

RankedHanging now has a non-configurable minimum rank of 2.