Replace Line node with BasicBlock, fix paragraph
Fix paragraph transformer: add a paragraph if TextBlock is not the only
sibling. Previously, we had to group Lines based on blank lines and
surrounding block elements.

Remove Line node as it does not represent a line anymore and replace it
with a BasicBlock which is just a Block node that can contain Inline
children.