Parse autolink URLs
Autolink URLs have either a "http://" or "https://" prefix. As such, any
such prefix is an autolink, irregardless of the following content. The
length of the autolink is determined with a few checks:
- domain validation (precondition for the following checks)
- trailing puncutation check (not part of autolink)
- unmatched trailing parentheses (not part of autolink)

A valid domain is a sequence of segments of alphanumeric characters,
underscores "_", and hypens "-" separated by periods ".". No underscores
may be present in the last two segments of the domain.