Better define element precedence
Match the most specific element-the one with the longest delimiter. If a
block and inline element have the same specificity, block element takes
precedence.

Before, if an inline element matched it was selected over any matching
block element. This was ok as inline elements always had longer
delimiters (except specific block elements which had delimiters that
made no sense as an inline delimiter).

Now, we can have blocks with the same root delimiter as an inline
element that take precedence over that inline element. E.g., we can have
block "-", inline "--", and block "---".