Add table recognizer
Table recognizer is a preprocessor which detects table delimiters and
transforms them into Touch element notation. For example, a "|" is a
table delimiter or rather a cell separator. When the table recognizer
encounters a "|" it outputs the content of the annotated cell as an
element Cell "&".

Example where "|" is a cell separator:
	a|b
Output where "<" is the row element and "&" the cell element:
	<&a
         &b

Currently, table recognizer only parsese inline elements. As such, it
knows not to parse "|" inside escaped elements but it doesn't know not
to parse "|" inside fenced elements. Similarly, it treats block
delimiters as inline content and outputs them inside cells rather than
before them.