Refactor parser tests part 1
Use golden files to test parser with the new node as it is hard to do so
in Go as we add children with .AppendChild() method instead of being
able to directly write them as before.

The parser tests are located in testdata/ directory and are grouped by
subdirectories like they were grouped before, e.g., TestLeaf tests are
in the testdata/leaf directory. Each test case is in its own .input
file. The wanted output is in a <name>.golden file. If any errors are
expected, they are placed in a <name>.error file. Each test subdirectory
must have an elements.json file that defines the parser elements for all
tests cases located in the directory.