Break up big printer test: complete
Three related printer tests fail: we don't print empty blocks and when
inserting newlines between lines treat them as if they were not there.
This causes an issue that we, between leaf and paragraph nodes, insert
a newline but this blocks would not have a newline in them if they were
not separated by the empty blocks. We would need to recognize this
blocks that shouldn't have a newline when broken up...