Put text contents into text nodes
Previously all nodes with verbatim content stored it in the Node's Value
field. Now, only the text node uses Value and others use the text node.
This was changed to be more consistent. For example, to render an
element we can always just render the children (except of course for the
text element). Or, searching for text now requires only checking the
text nodes.