parser: do not replace illegal characters
No longer replace non-first BOMs, NULLs, and illegal UTF-8 encodings
with the replacement character, but still keep reporting. Replacing
should not need to be the parsers domain.

Changing this will allow to use slices instead of needing to write
everything into a buffer (because we always needed to use the p.ch to
get replaced characters).