parser: use source slices instead of buffers
Before , we always needed to use p.ch because of illegal/invalid
character replacements. (If we used a source slice, we would not get the
replaced character.) In the previous commit we changed that, we removed
character replacements so now we can use the source bytes directly as
there is no point in writing everything into a buffer first anymore.

Also, keep newlines in text (before we replaced them with spaces).