Allow for elements to be disabled via config
Add bool field Disabled to config.Element that disables an element.
Disabled element is treated as if it were not at all present.

Disabled is used to disable default elements or elements from the
underlying config (in case multiple configs are supplied) as there is no
way to remove elements otherwise.