Add sourcemap to hold positional information
A sourcemap allows for mapping node's offset positions to a source file.
If a sourcemap is provided to the parser, the parser will mark line and
lead positions in it. Based on line positions, the sourcemap can tell us
the line and column positions just from an offset. And based on the lead
positions, it can tell us the subranges of lines an element spans.

Example:
	>
	 >

Before, we would have a hard time determening that the above walled
element spans (0:0-0:1) and (1:1-1:2) lines and columns (0-based). We
only stored the start and end offsets (0:4) which means that we would
count the offset 2 (space) as part of the element which it isn't.

Now, using the line and lead positions, the sourcemap can tell us
exactly which subranges of a line an element spans.