sourcemap: discren between walled and hanging
Sourcemap now marks blank leads which allows it to differentiate between
walled and hanging elements. Blank leads are the ' ' leads in parser.

A (walled):
	>
	 >

B (hanging):
	-
	 -

In both examples, the second line has lead columns: [0, 1].

Before, sourcemap had no other information and could not determine
whether the space at the start of the second line is in the element's
range or not. (It is in B but not in A.)

Now, sourcemap uses the blank markers to determine this. The blank
markers basically tell us whether the element is hanging and if spacing
is part of the element.